pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-argus_env"
version = "1.0.0"
description = "ARGUS — ML Research Integrity Environment for OpenEnv. Trains agents to detect missing baselines, cherry-picked results, and benchmark contamination in ML papers."
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    "openenv-core[core]>=0.2.2",
    # Inference dependencies
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m argus_env.server.app
server = "argus_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["argus_env", "argus_env.server"]
package-dir = { "argus_env" = ".", "argus_env.server" = "server" }